crashtesting: forum-mso-de-92960 GetEnglishSearchFontName varying results
depending on if it's called once, or twice on the input string.

The first converts from Full-Width Characters to equivalent lower ASCII
range with non letters retained. And then the second iteration would
convert to lower case ASCII with non letters filtered out. Presumably
the intention is the Full-Width case should get directly to the same
results as multiple calls.

Change-Id: Idba4ebe04c907c160ee53abf6d5551550da032dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135678
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>