sw_redlinehide: skip unnecessary updates when undoing redlined delete
When reproducing tdf#135976 and then Undo, an UAF crash happens here:

  assert(!pFrame->GetDrawObjs() || !pObjs->Contains(*pObj));

The pObjs was actually deleted and then re-created, because the pObj was
removed from the frame and added again to the same frame.

This is a bit unexpected, so prevent it by taking a shortcut in the
caller UpdateFramesForRemoveDeleteRedline() to insert a check that had
been removed in commit bd39942a4d8037d13d513f81177a90dccb1986e7.

If the rPam is inside a single node, the sw::RedlineUnDelText hint that
was sent to the SwTextFrame should be sufficient to update it and the
rest of the code in the loop that deals with newly split paragraph can
be skipped.

Change-Id: I5f36eb91bc20003887ee0bad03ea4a6e67135de9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135907
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>