Revert "simplify SharedString::operator=(&&)"
This reverts commit c1704b8645f9c6c551ad4a96af8511269ebe05e3.

Comment from llunak:
I'm possibly not remembering it exactly. But the point is that if you're going to destroy something, it's faster to destroy it outright rather than move it somewhere and destroy it later. If you google it you'll find posts on how implementing std::move() using std:swap() is generally a pessimisation.

Change-Id: Ice4234ad074cd894a25782e9c52b34c5acb26c8f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135880
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>