vcl: small optimization determining if any renderers are active
Currently we look at every renderer to see if it is paused. However, you
can think of this differently - instead of using a universal quantifier
we can actually use an existential quantifier - if even one renderer is
not paused, then we can say that not every renderer is paused - thus no
global pause. Hence switch to any_of(), which stops the loop at the
first instance of a non-paused renderer.

Change-Id: I3b35bc41e86432374e4bc5fae0a2927ec8cc2309
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136412
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>