tdf#149654 sw content controls: fix missing string before inline SDT from DOCX
The document had a block SDT and inside that, a run SDT. The content
that is before the run SDT but already inside the block SDT was lost.

To work incrementally, it was intentional that once commit
02a8799fdc949d1ddb205a0ed57bea6ff6fa8425 (sw content controls, date: add
DOCX import, 2022-05-30) changed most of run SDTs to content controls,
it left block SDTs (and cell/row SDTs, too) unchanged, so they are still
mapped to fields. What was forgotten is that m_pImpl->m_pSdtHelper in
DomainMapper is a shared state: once a run SDT starts, it assumes that
no non-run SDTs are in progress.

Fix the problem by explicitly checking for non-run SDTs before
PushSdt(), that keeps the separation (content control for run SDT,
fields for the rest) but fixes the lost content.

This is for plain text SDTs, but other types can be added if necessary.

Change-Id: I46b029a3a945d7416028aa196ac3160e6d96eae8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136524
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins