lok: allow shrinking too big widgets
This is followup for:
https://cgit.freedesktop.org/libreoffice/core/commit/?id=af407df46bad3df3f698a532b3e43f1e5084897f
Add sanity check to avoid potential crash in the Collabora Online server

nExtraSpace can be negative for shrinking the widget - this is correct.
For example listbox can have many entries, in GetOptimalSize it returns
size required to show all of them, then here we calculate how much of
that size we can allow to be used. When we have too many entries we use
negative nExtraSpace.

The original problem is not that we get negative number but case when we
get value that is much bigger than size of a widget and it becomes
invisible.

So this reverts hack which makes impossible to shrink widget what causes
problems with listboxes in tunneled dialogs. Also it setups better
size request for csv import dialog grid widget - so we don't paint it
as 10px rectangle - what was workarounded by the previous patches.

Change-Id: I478660b12cf0b6a783c704201800018e9849cc23
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136728
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mert Tumer <mert.tumer@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137451
Tested-by: Jenkins
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>