tdf#149816 qt a11y: Report position in screen coordinates
`QAccessibleInterface::rect()` is supposed to report the
geometry in screen coordinates [1].
`XAccessibleComponent::getLocationOnScreen` returns the
position in screen coordinates, so use that instead
of `XAccessibleComponent::getLocation`, which returns the
position relative to the parent.

With this in place, Accerciser correctly highlights
the area of the item curently selected in its treeview
of the LO a11y hierarchy.

[1] https://doc.qt.io/qt-6/qaccessibleinterface.html#rect

Change-Id: Iefb1e8792cd3ee720feafc81744e5e6ed6a949c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136781
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>