tdf#149978 sw: ODF import: fix nondeterministic automatic styles
The problem is that in SwpHints::TryInsertHint() there is a check for
IsInXMLImport() that is presumably an optimization to avoid the
potentially expensive call to BuildPortions().

While LO would only produce 1 text:span referencing an automatic style
around any given character content, this is not required by ODF, and so
other producers may legitimately produce such nested text:span elements.

Unfortunately the current SwpHints::Insert() isn't deterministic, the
RES_TXTATR_AUTOFMT with same start/end will be compared by address in
CompareSwpHtStart() (whereas RES_TXTATR_CHARFMT has a sort number for
this), so the result is going to be a random order.

Change-Id: Id62a7ff5fb85dbe42b7e1a27b0d8b36b74cf1100
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137033
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>