tdf#139759 writerfilter: import comment highlight
Although it is a highlight, LO doesn't have such a silly
duplicate thing in the editeng code. So just map this as
normal char background instead of char highlight.

As of LO 7.x, we default to exporting as char background anyway,
so highlight is on the way out in LO.

P.S. Highlight is one of 17-ish colors. It is the background
button in the Char panel in MS Word. Background is on Word's
Para panel (even though it is a character property),
and so it can fairly easily be removed in MS Word.

I didn't add a unit test because it doesn't round-trip,
and an import-only test is basically useless.
I'm having a surprisingly hard time finding why it doesn't export.

Change-Id: Iad279b503b9f307994f1d9b96e7d984d6d8b44fc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137036
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>