tdf#149548 sw: don't rely on binary search in SplitRedline()
The problem is that for this bugdoc overlapping redlines are created by
writerfilter, and so this one survives the split in SwDoc::TextToTable()
when creating a cell on node 146 to 149:

$125 = (SwRangeRedline) {
  <SwPaM> = SwPaM = {
    point = SwPosition (node 155, offset 156),
    mark = SwPosition (node 146, offset 0)
  },

It's not found in GetRedline(), as that relies on the redlines never
overlapping, so isn't split at end of 149.

Then when copying to the clipboard, all redlines are deleted and this
deletes cell start and end nodes, while the cell's SwTableBox is still
referenced elsewhere.

(reportedly somehow a regression from
commit 6949cc134aff305eb5573812f5fafc263e71963c)

Change-Id: I6466e6777cd83c7625381f7049f30c8a2f487af1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137110
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>