tdf#64242 sw table: improve Optimal Column Width function logic
This affects both Writer's minimze and optimize table functions.

Instead of basing equal-width on the current table size,
allow each column to have access to an equal portion
of the full table size (in case that it can grow wider).

This will help to maintain symmetry - one of the two primary
intentions of these functions.

[This is irrelevant for Draw - since the table size does not grow.]

Change-Id: I0e8c499b66c538f06df52eb74374bad3b690e8d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137120
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>