sw content control, picture: allow replacing via the keyboard
It was not possible to replace a picture placeholder with an actual
bitmap without using the mouse, which breaks accessibility.

The mouse handling code was added in commit
3a97719b7f50a782c3ff7a5d6059935f38feebdc (sw content controls, picture:
replace placeholder image on click, 2022-05-18), which already split the
functionality between SwEditWin::MouseButtonUp() and
SwWrtShell::GotoContentControl().

Fix the problem by reusing the shared SwWrtShell::GotoContentControl()
and extending SwEditWin::KeyInput() to allow doing the same with the
keyboard. This way the scenario when Shift-F4 selects the frame and the
user presses Enter now triggers the filepicker.

A possible future improvement would be to also handle Enter similarly
when the text cursor is inside a picture content control, but that's not
implemented here.

Change-Id: I756395d3811e3f4dfdce698751c4de13a0d49729
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137168
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins