ScriptForge - (SF_Dialog) trap errors caused by dialog events
A dialog is started with the Execute() method.

During its life span user scripts may be executed
as a consequence of mouse clicks etc.

When those scripts contain errors and they are not
trapped;
- in the actual situation, the error is detected by
  ScriptForge and considered as an internal error
- this new commit considers the error as a user error.

The help pages should be completed to promote the
use of errors trapping inside scripts triggered
by events.

Change-Id: Iba3eefc75d305c17e6e1d9d416496882e0f816bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137203
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins