sw content control, date: allow selecting via the keyboard
It was not possible to select the date of a date content control via
keyboard.

This meant only the mouse handling codepath needed implementing, but it
broke accessibility.

SwContentControl::ShouldOpenPopup() already knows how to handle
dropdowns via the keyboard, extend that to handle date pickers as well.

The appearing vcl Calendar / GtkCalendar appears to support keyboard
navigation already.

Change-Id: Ic36419f69cbcdba1a9069332184966cb4b347c49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137213
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins