tdf#150017 sw: allow editing bookmark text in Insert->Bookmarks dialog
Enable editing the text of the bookmark if it's "short" enough and
doesn't span multiple paragraphs; Replace the text when finished.

This will delete any footnotes or text fields in the bookmark text,
which isn't ideal but there's Undo.

Double-clicking the column via property "editable" only seems to work
with native GTK widgets, so provide an additional "Edit Text" button to
start the editing.  The button requires commit
2a0ea7b0220ad1ffb27d78187df6f7dd9ec69339 to edit the correct column
with VCL widgets, and it doesn't work yet with GTK widgets (but with GTK
double-clicking works).

Change-Id: If9e8a148b039889924e4870b2e9bbe977c977ce9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137355
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>