lok: initialize the load-language
We need to use the load-language for
saving the document. This is to avoid
using the language of the view that is
issuing the save, which causes all sorts
of issues (language-translation related).
This logic was implemented in 2657a1d35de4
but for some reason the load-language
wasn't set.

This patch sets the load-language.

Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>
Change-Id: I0cd1574d48a99e6ee84bacf75fc0d9ebe26e526b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137391
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 6024c8bc7ae9111a328d432120f2fb1b569f07f6)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140362
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>