vcl: fix assert fail on invalid user input in the vcl::WindowData ctor
This seems to be a problem since
45d0921daf1992c5a5586c1d65e81dd47a9f216e (VCL add vcl::WindowPosSize
abstract class, 2022-06-02), which started asserting that the window
size is not negative. Which is fine, but then callers like the
vcl::WindowData ctor has to be adapted, since they get user input.

My user profile had a string like:

"1056298152,1049249376,-1906262000,-1906261808;5;0,0,0,0;"

passed to that ctor, so soffice failed to start up.

Just ignore the broken user input at the call-site.

Change-Id: I942b0f2dfc1a07f9514c120761e100691f6d1a59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137357
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins