tdf#150064 Keep a11y child order intact
    commit bbc81e88d17acda14a07734c933d9b6bc6d32651
    Date:   Wed Jun 29 19:47:20 2022 +0200

        tdf#137544 reduce cost of ChildrenManagerImpl::Update

had added sorting based on memory addresses of the
corresponding shapes, but the order of the
children in `maVisibleChildren` is expected
to reflect the order of the accessible children
(i.e. based on the child index of the corresponding
shapes' a11y objects), since items are accessed by
index (s. e.g. `ChildrenManagerImpl::GetChild`).

Since the `ChildDescriptor`'s `mxAccessibleShape`
reference can be empty, its child index also cannot
be used for sorting instead.

To prevent the a11y tree from becoming unstable/random,
don't reorder/sort `maVisibleChildren`, but
allocate a helper vector holding pointers to the items
in the real vector, and iterate over that one instead.

This also moves identification of the elements
that are only in the old, but no longer the
new vector to `ChildrenManagerImpl::MergeAccessibilityInformation`,
where the sorted vector is availabe, and returns
a vector of obsolete children that is then
passed to
`ChildrenManagerImpl::RemoveNonVisibleChildren`,
instead of doing the comparison of the old/new
vector there.

Change-Id: Ie449f76f1b98ffe8e85ca28e938b11d726086721
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137622
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Colomban Wendling <cwendling@hypra.fr>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>