sc, out of order undo: allow a subset of a non-empty redo list
This is the calc analogue of
    commit 7100dc92548baac685e3cba6fbf09fa631f561b9
    sw, out of order undo: allow a subset of a non-empty redo list

Specifically, we used to not allow out of order undo at all if the redo
list was non-empty. This relaxes that condition a bit. Out of order undo
is OK with a non-empty redo list, in case all undo actions in the redo
list are either

1) owned by the current view or

2) independent from the undo action to be executed

I.e. if view1 has lots of type undo actions and an view2 adds a
single type undo action on top of it, then allow view 1 to execute
multiple of its typing undo actions, not just a single one.

Change-Id: Ib7ab88e9ebbefce267d3c7d567debd1cd98d7d6d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137628
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>