tdf#150034 very slow opening pathological chart
This chart has a ridiculously long legend, which makes the loop that
tries to shorten legends take an extremely long time.

Make this loop faster, at the risk of sometimes getting a slightly wrong
answer.
Ideally we could speed up the layout down in ImpEditEngine, but that is
a much bigger task.

Change-Id: I7cb337c674515bac13e4b11c3b0fabb94aed6865
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137677
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>