sw: always show "tabs" tab in the paragraph dialog for multiple paragraphs
In case multiple paragraphs were selected and the various paragraphs had different left/right margins in their item sets, we used to hide the "tabs" tab page on the paragraph dialog.

This had the benefit that we avoided presenting information that is not entirely true for all paragraphs, but had the downside that it was not possible to set tab stops for multiple paragraphs when at least one of the paragraph had different left/right margins compared to the reset of the paragraphs. This required the user to "hunt" for the offending paragraph(s). The ability to always have the possibility to set tab stops seems to provide more benefit than the cost of presenting a partially confusing "tabs" tab page when reading the doc model.

Fix the problem by relaxing the requirements and only hiding the "tabs" tab page in the writer/web page, show the tab page is now always shown for normal documents.

This change makes no effort to show the common tab stops of all selected paragraphs on the UI, which would probably make sense, but can be done in a follow-up change.

Signed-off-by: Pranam Lashkari <lpranam@collabora.com>
Change-Id: Ida61e8878382988a7470d156ded187d21bcc9d91
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137707
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 36e90d98df9fd0b9aef12ef11ade358676fe2963)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137718
Tested-by: Jenkins