wina11y: Drop AccObjectManagerAgent::UpdateLocation
That static method doesn't do anything.

`_IMPL_WIN` is not defined, so the

    #ifdef _IMPL_WIN
       if( pWinManager )
           pWinManager->SetLocation( pXAcc, top, left, width, height );
    #endif

in the method is unused.

Since `CMAccessible::accLocation` already retrieves the
current location from the underlying `XAccessible` on demand,
I don't see a need to update the location manually
in the first place, so there should be no need to enable that,
code, just drop it.

Also drop `AccObjectWinManager::SetLocation` which was
unused since `AccObjectManagerAgent::UpdateLocation`
(s. above) would have been the only call site.

Change-Id: I04d0230d3599466aaa92082caba54da22a3b1a28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/137857
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>